Improved Weather API

+++ DiscordTest/DataTypes/weather.cs
@@ -19,7 +19,7 @@ namespace DiscordTest.DataType
         public class Main
         {
             public double temp { get; set; }
-            public int pressure { get; set; }
+            public double pressure { get; set; }
             public int humidity { get; set; }
             public double temp_min { get; set; }
             public double temp_max { get; set; }
